agents/proxy/requirements.txt
fastapi
uvicorn
google-auth
google-auth-httplib2
requests